public_scrapper/public_scrapper/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
Automat==25.4.16
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
constantly==23.10.4
cryptography==46.0.6
cssselect==1.4.0
defusedxml==0.7.1
dnspython==2.8.0
dotenv==0.9.9
fastapi==0.135.3
filelock==3.25.2
greenlet==3.3.2
h11==0.16.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
jmespath==1.1.0
lxml==6.0.2
packaging==26.0
parsel==1.11.0
playwright==1.58.0
Protego==0.6.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyDispatcher==2.0.7
pyee==13.0.1
pymongo==4.16.0
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-whois==0.9.6
queuelib==1.9.0
requests==2.33.0
requests-file==3.0.1
Scrapy==2.14.2
scrapy-playwright==0.0.46
scrapy-user-agents==0.1.1
service-identity==24.2.0
six==1.17.0
starlette==1.0.0
tldextract==5.3.1
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==202603
urllib3==2.6.3
user-agents==2.2.0
uvicorn==0.42.0
w3lib==2.4.1
zope.interface==8.2
